Lib name and fixing an issue
I decided to change the name of the library to Mainframe mostly because the previous name could lead to people thinking of an open range of console/terminal applications and the new name reflects exactly the target kind of application it is aimed at producing.

Second, removed the (public) hability to provide the map contents in the constructor because it could lead to bugs if an application tries to load a map from remote contents.