Reverted field info to internals
During the initial development of the framework, fields used a single class independently of field type but it was reformed to separate each field type in a class and allow for new user defined types.

In the same vision of focusing on the least effort, I found it was best to remove the possibility to add custom field types and reverted to a single field type for all fields so that there is no need for users to waste time seeking a possible component field for development since there will be none.

I also took advantage of this move to hide internal details of fields and rework how fields are referenced by applications, now by name.