Restricted arrow navigation to fields
This is something I noticed after the initial implementation. A behaviour I copied from the terminal app my wife uses allows for the cursor to roam free on the screen no matter where fields are. I don't know of a reason for that (except maybe to make it behave like a mouse pointer, which is actually ignored). By pressing the TAB key the user can only move through editable/focusable fields. This restricted navigating using the arrow keys to the region of known field.

Also took advantage to change how information about a field is set using flags that can be combined as current and future fields require.