Fix handling of window size
Window size, by the initial code, would resize the command prompt to this size only and restore it once the application is finished. Maybe it is not actually interesting to resize the window at all times but certain situations could make an application output the map contents beyond the known limits.

After some thoughts on how to implement cropping larger maps, it occurred to me that, no matter how, doing so might cause some fields to be removed from the screen too and thus those fields would also have to be ignored while navigating between them. This would demand large changes to the entire framework.

The solution will be to raise an exception if the application calls Show() on a map that is larger than the application screen can cope with. Also, it would be unfeasible to consider the size of the application screen the same as the console at all times because this would make this behaviour erratic. Thus, unless specified by the developer, the default application screen will be 80 chars per 24 lines.