Enhancing colour support
In this commit I reduced the set of colour options for the developer and also removed the possibility of using different colours for fields. This is consistent with the principle of not wasting too much time with the UI.

Also changed the way IConsole works having it handle colours and preparing it to handle best outputing of content to the screen.