Redistributed code in fields
This has already subject to a previous change but it was reversed in this commit in order to better distribute code handling according to field type. The field classes are still not visible to the developer and new field classes cannot be registered by the developer without changing the library code.

More tests are pending.