Extracted indirectes from SystemConsole
IConsole interface was originally created such that I could abstract the System.Console methods in order to create a mock console to test the framework without resorting to viewing its results on screen. But previous changes added a lot of processing that could not be tested. This change reverted that so that SystemConsole methods only call directly System.Console methods.

As a result, a new class was created wrap around the IConsole instance and provide the methods that would not be tested if directly under SystemConsole. That also elevated the total test coverage.