Fixed rendering in Windows Terminal
There was an issue regarding rendering of maps on Windows Terminal when the last cursor left position considered was equals to the width of the screen (should actually be width - 1).

Also was found and fixed an issue where pressing the END key on a field would have a different behaviour depending on the cursor position within the field.